requirements.txt
leonardo-admin>=2015.7.0
django-material==0.4.1